// pixel-art/src/store/canvas/enum.ts
export enum MutationsTypes {
  SET_PAINT_MODE = "SET_PAINT_MODE",
  CREATE_LAYER = "CREATE_LAYER",
  CREATE_TEMP_LAYER = "CREATE_TEMP_LAYER",
  CREATE_PAGE = "CREATE_PAGE",
  SET_CANVASCTX = "SET_CANVASCTX",
  SET_COLOR = "SET_COLOR",
  SET_START_POINT = "SET_START_POINT",
  SET_END_POINT = "SET_END_POINT",
  SET_LASET_START_POINT = "SET_LASET_START_POINT",
  SET_LASET_END_POINT = "SET_LASET_END_POINT"
}

export enum ActionsTypes {
  SET_PAINT_MODE = "SET_PAINT_MODE",
  CREATE_LAYER = "CREATE_LAYER",
  CREATE_TEMP_LAYER = "CREATE_TEMP_LAYER",
  CREATE_PAGE = "CREATE_PAGE",
  SET_CANVASCTX = "SET_CANVASCTX",
  SET_COLOR = "SET_COLOR",
  SET_START_POINT = "SET_START_POINT",
  SET_END_POINT = "SET_END_POINT",
  SET_LASET_START_POINT = "SET_LASET_START_POINT",
  SET_LASET_END_POINT = "SET_LASET_END_POINT"
}